pyproject.toml
[tool.poetry]
name = "realmate-challenge"
version = "0.1.0"
description = ""
authors = ["viniciusfelippetti <felippettivinicius@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.2.1"
djangorestframework = "^3.16.0"
psycopg2-binary = "^2.9.10"
drf-yasg = "^1.21.10"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
